requirements.txt
Flask
PyMuPDF
requests
python-dotenv